Added packages: ruibot, ruibot.tasks
Let's:
* Package all classes under ruibot package
* Under ruibot package, create tasks package to add all Task classes: Task, ToDo, Deadline, Event
