+++ data/ruibot.txt
@@ -3,5 +3,5 @@
 [E] [ ] project meeting (from: Feb 03 2024, 6:00 am to: Feb 03 2024, 8:00 am)
 [D] [ ] return book (by: Feb 01 2024, 6:00 pm)
 [T] [ ] run 2
-[D] [ ] return book 2 (by: Feb 01 2024, 6:00 pm)
-[E] [X] project meeting 2 (from: Feb 03 2024, 6:00 am to: Feb 03 2024, 8:00 am)
+[D] [X] return book 2 (by: Feb 01 2024, 6:00 pm)
+[E] [ ] morning meeting (from: Feb 03 2024, 6:00 am to: Feb 03 2024, 8:00 am)
