+++ src/main/java/ruibot/EmptyTaskException.java
@@ -1,3 +1,5 @@
+package ruibot;
+
 public class EmptyTaskException extends Exception {
     public EmptyTaskException() {
         super("Task cannot be empty.");
