+++ src/main/java/ruibot/Parser.java
@@ -1,3 +1,5 @@
+package ruibot;
+
 public class Parser {
 
     public static String read(String input) {
