+++ src/main/java/ruibot/TaskList.java
@@ -1,4 +1,9 @@
-import java.lang.reflect.Array;
+package ruibot;
+
+import ruibot.tasks.Task;
+import ruibot.tasks.ToDo;
+import ruibot.tasks.Deadline;
+import ruibot.tasks.Event;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
