+++ src/main/java/ruibot/WrongInputException.java
@@ -1,3 +1,5 @@
+package ruibot;
+
 public class WrongInputException extends Exception {
     public WrongInputException() {
         super("I don't understand what that means.");
