+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
-Hello! I'm RuiBot
+Hello! I'm ruibot.RuiBot
 What can I do for you?
 ____________________________________________________________
